Task/adb 371

+++ app/src/main/java/com/example/movieme/domain/usecase/MovieUseCase.kt
@@ -9,6 +9,6 @@ class MovieUseCase @Inject constructor(private val movieRepository: MovieReposit
     ) = movieRepository.getMovieList(page = page)
 
     suspend fun getMovie(
-        movieId: Int,
+        movieId: Long,
     ) = movieRepository.getMovie(movieId = movieId)
 }
